Xpruning-blocks-retained should accept an integer in the config. (#440)

+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/worldstate/PrunerConfiguration.java
@@ -15,14 +15,14 @@
 package org.hyperledger.besu.ethereum.worldstate;
 
 public class PrunerConfiguration {
-  public static final long DEFAULT_PRUNING_BLOCKS_RETAINED = 1024;
-  public static final long DEFAULT_PRUNING_BLOCK_CONFIRMATIONS = 10;
+  public static final int DEFAULT_PRUNING_BLOCKS_RETAINED = 1024;
+  public static final int DEFAULT_PRUNING_BLOCK_CONFIRMATIONS = 10;
 
-  private final long blocksRetainedBeforeSweeping;
-  private final long blockConfirmationsBeforeMarking;
+  private final int blocksRetainedBeforeSweeping;
+  private final int blockConfirmationsBeforeMarking;
 
   public PrunerConfiguration(
-      final long blockConfirmationsBeforeMarking, final long blocksRetainedBeforeSweeping) {
+      final int blockConfirmationsBeforeMarking, final int blocksRetainedBeforeSweeping) {
     this.blockConfirmationsBeforeMarking = blockConfirmationsBeforeMarking;
     this.blocksRetainedBeforeSweeping = blocksRetainedBeforeSweeping;
   }
@@ -32,11 +32,11 @@ public class PrunerConfiguration {
         DEFAULT_PRUNING_BLOCK_CONFIRMATIONS, DEFAULT_PRUNING_BLOCKS_RETAINED);
   }
 
-  public long getBlocksRetained() {
+  public int getBlocksRetained() {
     return blocksRetainedBeforeSweeping;
   }
 
-  public long getBlockConfirmations() {
+  public int getBlockConfirmations() {
     return blockConfirmationsBeforeMarking;
   }
 }
